refactor: simplify join subcommand
